tests: Do not depend on useradd in TEST-19-CGROUP.delegate
The current form allows failure of useradd but then runs subsequent test
anyway. Bail out and report a skipped (sub)test.

+++ test/units/testsuite-19.delegate.sh
@@ -56,7 +56,10 @@ function test_controllers() {
 }
 
 test_scope_unpriv_delegation() {
-    useradd test ||:
+    if ! useradd test ; then
+        echo "Skipping TEST-19-DELEGATE unprivileged delegation test, can't create users" >&2
+        return
+    fi
     trap "userdel -r test" RETURN
     systemd-run --uid=test \
                 --property="User=test" \
